# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: DairyCare Foundation API
  version: 2.1.0-phase2b
  description: Phase 1 foundation plus the Phase 2A animal registry and Phase 2B online animal movement workflow.
servers:
  - url: http://127.0.0.1:8000/api/v1
security:
  - opaqueSession: []
paths:
...
  /animals/{animal}/movements:
    parameters:
      - $ref: '#/components/parameters/AnimalId'
    get:
      summary: List authorized movement history for an animal
      operationId: listAnimalMovements
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - name: 'filter[status]'
          in: query
          schema: { $ref: '#/components/schemas/AnimalMovementStatus' }
        - name: 'filter[search]'
          in: query
          schema: { type: string, maxLength: 120 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [requested_effective_at, -requested_effective_at, created_at, -created_at]
      responses:
        '200': { $ref: '#/components/responses/SuccessResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    post:
      summary: Request an online animal location movement
      operationId: requestAnimalMovement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnimalMovementCreate' }
      responses:
        '201': { $ref: '#/components/responses/SuccessResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
...
  /animal-movements/{movement}/approve:
    parameters:
      - $ref: '#/components/parameters/AnimalMovementId'
    post:
      summary: Approve and atomically apply a pending animal movement
      operationId: approveAnimalMovement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnimalMovementDecision' }
      responses:
        '200': { $ref: '#/components/responses/SuccessResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '412': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /animal-movements/{movement}/reject:
    parameters:
      - $ref: '#/components/parameters/AnimalMovementId'
    post:
      summary: Reject a pending animal movement
      operationId: rejectAnimalMovement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnimalMovementReasonDecision' }
      responses:
        '200': { $ref: '#/components/responses/SuccessResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '412': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /animal-movements/{movement}/cancel:
    parameters:
      - $ref: '#/components/parameters/AnimalMovementId'
    post:
      summary: Cancel a pending animal movement
      operationId: cancelAnimalMovement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnimalMovementReasonDecision' }
      responses:
        '200': { $ref: '#/components/responses/SuccessResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
        '412': { $ref: '#/components/responses/ErrorResponse' }
        '422': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    opaqueSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
  parameters:
...
    AnimalId: { name: animal, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    AnimalMovementId: { name: movement, in: path, required: true, schema: { $ref: '#/components/schemas/Uuid' } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string, maxLength: 160 } }
    PageSize: { name: 'page[size]', in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
...
  schemas:
    Uuid: { type: string, format: uuid }
...
    AnimalMovementStatus: { type: string, enum: [pending, approved, rejected, cancelled] }
...
    AnimalMovementCreate:
      type: object
      required:
        - source_farm_id
        - source_shed_id
        - destination_farm_id
        - destination_shed_id
        - requested_effective_at
        - reason
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        source_farm_id: { $ref: '#/components/schemas/Uuid' }
        source_shed_id: { $ref: '#/components/schemas/Uuid' }
        source_animal_group_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        destination_farm_id: { $ref: '#/components/schemas/Uuid' }
        destination_shed_id: { $ref: '#/components/schemas/Uuid' }
        destination_animal_group_id:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
        requested_effective_at:
          type: string
          format: date-time
          description: Requested business time, which cannot be in the future in Phase 2B.
        reason: { type: string, maxLength: 255 }
        notes: { type: [string, 'null'], maxLength: 5000 }
    AnimalMovementDecision:
      type: object
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
    AnimalMovementReasonDecision:
      type: object
      required: [version, reason]
      properties:
        version: { type: integer, minimum: 1 }
        reason: { type: string, maxLength: 1000 }
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          type: object
          required: [request_id]
          properties:
            request_id: { $ref: '#/components/schemas/Uuid' }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields, details, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: true }
            details: { type: object, additionalProperties: true }
            request_id: { $ref: '#/components/schemas/Uuid' }
...
  responses:
    SuccessResponse:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    ErrorResponse:
      description: Stable API error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }